# backend/src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Authentication API",
    "version": "1.0.0",
    "description": "API documentation for user authentication."
  },
  "basePath": "/",
  "paths": {
...
    "/employee/{id}": {
      "delete": {
        "summary": "Delete an employee by ID",
        "description": "Deletes an employee by their unique ID.",
        "tags": [
          "Employee"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "The ID of the employee to delete."
          }
        ],
        "responses": {
          "200": {
            "description": "Employee deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Employee not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Employee not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error occurred during deletion",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Internal server error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get an employee by ID",
        "description": "Retrieves an employee's information by their unique ID.",
        "tags": [
          "Employee"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "The ID of the employee to retrieve."
          }
        ],
        "responses": {
          "200": {
            "description": "Employee retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "description": "The ID of the employee."
                    },
                    "email": {
                      "type": "string",
                      "description": "The email of the employee."
                    },
                    "firstName": {
                      "type": "string",
                      "description": "The first name of the employee."
                    },
                    "lastName": {
                      "type": "string",
                      "description": "The last name of the employee."
                    },
                    "permissions": {
                      "type": "array",
                      "description": "List of permissions for the employee.",
                      "items": {
                        "type": "string"
                      }
                    },
                    "phoneNumber": {
                      "type": "string",
                      "description": "The phone number of the employee."
                    },
                    "timeRecords": {
                      "type": "array",
                      "description": "List of time records for the employee.",
                      "items": {
                        "type": "object"
                      }
                    },
                    "role": {
                      "type": "string",
                      "description": "The role of the employee."
                    }
                  }
                },
                "example": {
                  "id": 1,
                  "email": "john@example.com"
                }
              }
            }
          },
          "404": {
            "description": "Employee not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Employee not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error occurred during retrieval",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Internal server error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update an employee by ID",
        "description": "Updates an employee's information by their unique ID.",
        "tags": [
          "Employee"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "The ID of the employee to update."
          },
          {
            "in": "body",
            "name": "employee",
            "required": true,
            "description": "The updated employee data.",
            "schema": {
              "type": "object",
              "properties": null
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employee updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "description": "The ID of the updated employee."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Employee not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Employee not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error occurred during update",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
    "/employees/{id}/time-records": {
      "get": {
        "summary": "Get all time records for an employee",
        "description": "Retrieves all time records for a specific employee based on their ID.",
        "tags": [
          "Time Records"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "The ID of the employee whose time records are to be retrieved."
          }
        ],
        "responses": {
          "200": {
            "description": "Time records retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "timeRecords": {
                      "type": "array",
                      "items": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid id parameter",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Invalid id parameter"
                }
              }
            }
          },
          "404": {
            "description": "Employee not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Employee not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error occurred during retrieval",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}